test change

+++ back-end/src/utils/codeFileManager.ts
@@ -1,7 +1,9 @@
-import fs, { readdirSync } from "fs";
+import fs, { existsSync, readdirSync } from "fs";
 import path, { join } from "path";
 import { v4 as uuidv4 } from "uuid";
 import {
+  PATH_TO_CODE_FOLDER,
+  PATH_TO_OUTPUT_FOLDER,
   SUPPORTED_LANGUAGES_EXTENSIONS,
   SUPPORTED_LANGUAGES_OUTPUT_EXTENSIONS,
 } from "../config/constants";
@@ -9,15 +11,17 @@ import { TLanguage } from "./validationSchema";
 
 export const createCodeFile = (code: string, language: TLanguage) => {
   const codeId = uuidv4();
-  let files = readdirSync(join(process.cwd(), "/usr/code"));
+  console.log(__dirname);
+
+  let files = readdirSync(join(__dirname, PATH_TO_CODE_FOLDER));
   console.log(files);
 
   const filePath = path.join(
-    process.cwd(),
-    `/usr/code/${codeId}.${SUPPORTED_LANGUAGES_EXTENSIONS[language]}`
+    __dirname,
+    `${PATH_TO_CODE_FOLDER}/${codeId}.${SUPPORTED_LANGUAGES_EXTENSIONS[language]}`
   );
   fs.writeFileSync(filePath, code);
-  files = readdirSync(join(process.cwd(), "/usr/code"));
+  files = readdirSync(join(__dirname, PATH_TO_CODE_FOLDER));
   console.log(files);
 
   return { codeId };
@@ -25,16 +29,20 @@ export const createCodeFile = (code: string, language: TLanguage) => {
 
 export const removeCodeFile = (codeId: string, language: TLanguage) => {
   const codeFilePath = path.join(
-    process.cwd(),
-    `/usr/code/${codeId}.${SUPPORTED_LANGUAGES_EXTENSIONS[language]}`
+    __dirname,
+    `${PATH_TO_CODE_FOLDER}/${codeId}.${SUPPORTED_LANGUAGES_EXTENSIONS[language]}`
   );
   fs.unlinkSync(codeFilePath);
 
-  if (SUPPORTED_LANGUAGES_OUTPUT_EXTENSIONS[language]) {
-    const outputFilePath = path.join(
-      process.cwd(),
-      `/usr/output/${codeId}.${SUPPORTED_LANGUAGES_OUTPUT_EXTENSIONS[language]}`
-    );
+  const outputFilePath = path.join(
+    __dirname,
+    `${PATH_TO_OUTPUT_FOLDER}/${codeId}.${SUPPORTED_LANGUAGES_OUTPUT_EXTENSIONS[language]}`
+  );
+
+  if (
+    SUPPORTED_LANGUAGES_OUTPUT_EXTENSIONS[language] &&
+    existsSync(outputFilePath)
+  ) {
     fs.unlinkSync(outputFilePath);
   }
 };
